backend/requirements.txt
amqp==2.4.0
argh==0.26.2
autopep8==1.4.3
billiard==3.5.0.5
boto3==1.9.78
botocore==1.12.78
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.5
django-celery-beat==1.4.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-timezone-field==3.0
djangorestframework==3.9.0
docutils==0.14
forex-python==1.3
idna==2.8
jmespath==0.9.3
json2html==1.2.1
kombu==4.2.2.post1
pathtools==0.1.2
pika==0.12.0
Pillow==5.4.1
pkg-resources==0.0.0
pycodestyle==2.4.0
pycparser==2.19
python-crontab==2.3.6
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.9
PyYAML==3.13
requests==2.21.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
